requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.9
einops>=0.7
albumentations>=1.4
opencv-python
scikit-image
scipy
numpy
pandas
scikit-learn
torchmetrics
matplotlib
tabulate
fvcore
ptflops
pyyaml
